Change usage of cookies
Initially we did not set any cookies or local storage. For making it possible to login to our services, we used the browsers session storage exclusively.

Because session storage is only accessible in the same tab, users had to frequently log in again to our services.

To make this more convenient, we will implement a way to stay logged in for longer and across browser tabs. To do so, we have to use either cookies or local storage.

We will still not set any third-party cookies. But to make our in-house analytics more accurate, we might decide to set a cookie.